+++ src/jsonSchema/jsonSchema.numeric-bounds.ts
